+++ packages/core/src/manglers/__tests__/html-attributes.test.ts
@@ -683,6 +683,8 @@ suite("HTML Attribute Mangler", function() {
   suite("HTML (style attribute)", function() {
     const embedDeclarationsInStyle = embedAttributeValue("style");
 
+    const varyDeclarationSpacing = varySpacing([":", ",", ";"]);
+
     type TestInstance = {
       readonly name: string;
       factory(idBefore: string, idAfter: string): TestCase[];
@@ -700,15 +702,6 @@ suite("HTML Attribute Mangler", function() {
         ],
       },
       {
-        name: "attribute usage (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ")"], {
-            input: `content:attr(${before});`,
-            expected: `content:attr(${after});`,
-          }),
-        ],
-      },
-      {
         name: "attribute usage with default",
         factory: (before: string, after: string): TestCase[] => [
           ...CSS_VALUES_NO_STRINGS
@@ -716,15 +709,6 @@ suite("HTML Attribute Mangler", function() {
               input: `content:attr(${before},${value});`,
               expected: `content:attr(${after},${value});`,
             })),
-        ],
-      },
-      {
-        name: "attribute usage with default (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ","], {
-            input: `content:attr(${before},42);`,
-            expected: `content:attr(${after},42);`,
-          }),
         ],
       },
       {
@@ -738,36 +722,33 @@ suite("HTML Attribute Mangler", function() {
         ],
       },
       {
-        name: "attribute usage with type/unit (spacing)",
+        name: "attribute usage with default and type/unit (different properties)",
         factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ")"], {
-            input: `content:attr(${before} px);`,
-            expected: `content:attr(${after} px);`,
-          }),
+          ...CSS_PROPERTIES
+            .map((property: string): TestCase => ({
+              input: `${property}:attr(${before} px,42);`,
+              expected: `${property}:attr(${after} px,42);`,
+            })),
         ],
       },
       {
-        name: "attribute usage with default and type/unit",
+        name: "attribute usage with default and type/unit (different types/units)",
         factory: (before: string, after: string): TestCase[] => [
           ...TYPE_OR_UNITS
             .map((typeOrUnit: string): TestCase => ({
               input: `content:attr(${before} ${typeOrUnit},42);`,
               expected: `content:attr(${after} ${typeOrUnit},42);`,
             })),
+        ],
+      },
+      {
+        name: "attribute usage with default and type/unit (different values)",
+        factory: (before: string, after: string): TestCase[] => [
           ...CSS_VALUES_NO_STRINGS
             .map((value: string): TestCase => ({
               input: `content:attr(${before} px,${value});`,
               expected: `content:attr(${after} px,${value});`,
             })),
-        ],
-      },
-      {
-        name: "attribute usage with default and type/unit (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ","], {
-            input: `content:attr(${before} px,42);`,
-            expected: `content:attr(${after} px,42);`,
-          }),
         ],
       },
     ];
@@ -845,6 +826,13 @@ suite("HTML Attribute Mangler", function() {
             .map(embedDeclarationsInStyle)
             .flatMap(varyAttributeSpacing)
             .flatMap(varyHtmlQuotes)
+            .flatMap(embedAttributesInTags),
+        },
+        {
+          name: `${name}, vary spacing`,
+          cases: factory("data-foobar", "data-a")
+            .map(embedDeclarationsInStyle)
+            .flatMap(varyDeclarationSpacing)
             .flatMap(embedAttributesInTags),
         },
         {
@@ -1058,6 +1046,28 @@ suite("HTML Attribute Mangler", function() {
           ],
         },
       ]);
+
+      if (!factory("data-foobar", "data-a")[0].input.includes(" ")) {
+        scenarios.push({
+          name: `${name} in an unquoted style attribute`,
+          cases: [
+            ...factory("data-foobar", "data-a")
+              .map(embedDeclarationsInStyle)
+              .flatMap((testCase: TestCase): TestCase[] => [
+                {
+                  input: testCase.input.replace(/"/g, ""),
+                  expected: testCase.expected.replace(/"/g, ""),
+                },
+                {
+                  input: testCase.input.replace(/("|;)/g, ""),
+                  expected: testCase.expected.replace(/("|;)/g, ""),
+                },
+              ])
+              .flatMap(varyAttributeSpacing)
+              .flatMap(embedAttributesInTags),
+          ],
+        });
+      }
     }
 
     run("html", scenarios);